Bug Fixes
Close:
Bug #289
Bug #290
Bug #291
Bug #292
Bug #293
Bug #294

+++ admin/pages/ias-brands-save.php
@@ -1,14 +1,16 @@
 <?php
-foreach ($_POST['selected'] as $brand) {
-	$region_original = json_decode( $wpdb->get_var( ias_fix_db_prefix("SELECT `brands` FROM `{{ias}}regions` WHERE `id` = 3") ) , true );
-	$region_updated = array();
-	foreach ($region_original as $b) {
-		if($b != $brand) {
-			array_push($region_updated,$b);
+if( isset($_POST['selected']) ) {
+	foreach ($_POST['selected'] as $brand) {
+		$region_original = json_decode( $wpdb->get_var( ias_fix_db_prefix("SELECT `brands` FROM `{{ias}}regions` WHERE `id` = 3") ) , true );
+		$region_updated = array();
+		foreach ($region_original as $b) {
+			if($b != $brand) {
+				array_push($region_updated,$b);
+			}
 		}
+		$wpdb->update( ias_fix_db_prefix('{{ias}}regions') , array('brands' => json_encode($region_updated)) , array('id' => 3) );
+		$wpdb->delete( ias_fix_db_prefix('{{ias}}brands') , array('id' => $brand));
 	}
-	$wpdb->update( ias_fix_db_prefix('{{ias}}regions') , array('brands' => json_encode($region_updated)) , array('id' => 3) );
-	$wpdb->delete( ias_fix_db_prefix('{{ias}}brands') , array('id' => $brand));
 }
 do_action('ias_delete_brand' , $_POST['selected'] );
 header("location: admin.php?page=ias-brands");
